# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "kasane",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.3.2"
  },
  "paths": {
...
    "/databases/{db_name}/tables/{table_name}": {
      "get": {
        "tags": [
          "Tables"
        ],
        "summary": "テーブル情報の取得",
        "description": "指定したテーブルの詳細情報を取得します。この操作はデータベースのRead以上の権限が必要です。",
        "operationId": "table_info",
        "parameters": [
          {
            "name": "db_name",
            "in": "path",
            "description": "データベース名",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example_database"
          },
          {
            "name": "table_name",
            "in": "path",
            "description": "テーブル名",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example_table"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableInfoResponse"
                }
              }
            }
          },
          "404": {
            "description": "テーブルまたはデータベースが存在しない"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Tables"
        ],
        "summary": "テーブルの削除",
        "description": "指定したテーブルを削除します。この操作はデータベースのManage以上の権限が必要です。",
        "operationId": "remove_table",
        "parameters": [
          {
            "name": "db_name",
            "in": "path",
            "description": "データベース名",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example_database"
          },
          {
            "name": "table_name",
            "in": "path",
            "description": "テーブル名",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example_table"
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": "テーブルが存在しない"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Tables"
        ],
        "summary": "テーブルの更新",
        "description": "テーブルの名前変更や値の制約を更新します。この操作はデータベースの Manage 権限が必要です。\n\n制約を変更する際、既存のデータが新しい制約に適合しているかを検証する挙動を `validate_existing_data` フラグで制御できます。\n- `true`（デフォルト）の場合、既存データが新しい制約に違反していればエラーとして更新を却下します。\n- `false` の場合、既存データのチェックをスキップして高速に変更を適用します（不整合なデータが残る可能性があります）。",
        "operationId": "table_update_handler",
        "parameters": [
          {
            "name": "db_name",
            "in": "path",
            "description": "データベース名",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "table_name",
            "in": "path",
            "description": "テーブル名",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTableRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "テーブルの更新に成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableSummary"
                }
              }
            }
          },
          "400": {
            "description": "リクエストが不正（パラメータエラー、または validate_existing_data=true の状態で既存データが新しい制約に違反した場合）"
          },
          "401": {
            "description": "認証エラー"
          },
          "403": {
            "description": "権限が不足している"
          },
          "404": {
            "description": "データベースまたはテーブルが見つからない"
          },
          "409": {
            "description": "同名のテーブルが既に存在する"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/databases/{db_name}/tables/{table_name}/data": {
      "put": {
        "tags": [
          "Data"
        ],
        "summary": "データの上書き・追加",
        "description": "指定した空間IDに対して、指定した値を書き込みます。\n既に値が存在する空間IDに対しては、その値を完全に上書きします。\n\nこの操作はデータベースのWrite以上の権限が必要です。",
        "operationId": "data_insert",
        "parameters": [
          {
            "name": "db_name",
            "in": "path",
            "description": "データベース名",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example_database"
          },
          {
            "name": "table_name",
            "in": "path",
            "description": "テーブル名",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example_table"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InsertDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": "テーブルが存在しない"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Data"
        ],
        "summary": "データの削除",
        "description": "指定した空間IDをテーブルからを削除します。この操作はデータベースのWrite以上の権限が必要です。",
        "operationId": "data_remove",
        "parameters": [
          {
            "name": "db_name",
            "in": "path",
            "description": "データベース名",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example_database"
          },
          {
            "name": "table_name",
            "in": "path",
            "description": "テーブル名",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example_table"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": "テーブルが存在しない"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Data"
        ],
        "summary": "データの部分追加",
        "description": "指定した空間IDに対して、指定した値を書き込みます。\n\n- **指定した空間IDがすでに存在する場合**: 既存のデータは維持されます。\n- **指定した空間IDが存在しない場合**: データが書き込まれます。\n\nこの操作はデータベースのWrite以上の権限が必要です。",
        "operationId": "data_upsert",
        "parameters": [
          {
            "name": "db_name",
            "in": "path",
            "description": "データベース名",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example_database"
          },
          {
            "name": "table_name",
            "in": "path",
            "description": "テーブル名",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "example_table"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InsertDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": "テーブルが存在しない"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FlexId": {
        "type": "object",
        "description": "高さ・経度・緯度・時間の4軸すべてを「ズームレベル + インデックス」で表す拡張時空間IDを表す型。\n時間を指定する場合は `t_zoomlevel` と `t_index` を必ず両方指定すること。\n両方省略した場合は「全時間」を表す。",
        "required": [
          "fZoomlevel",
          "fIndex",
          "xZoomlevel",
          "xIndex",
          "yZoomlevel",
          "yIndex"
        ],
        "properties": {
          "fZoomlevel": {
            "type": "integer",
            "format": "int32",
            "description": "高さ方向のズームレベル。",
            "example": 20,
            "maximum": 30,
            "minimum": 0
          },
          "fIndex": {
            "type": "integer",
            "format": "int32",
            "description": "高さ方向のインデックス。",
            "example": 0
          },
          "xZoomlevel": {
            "type": "integer",
            "format": "int32",
            "description": "経度方向のズームレベル。",
            "example": 20,
            "maximum": 30,
            "minimum": 0
          },
          "xIndex": {
            "type": "integer",
            "format": "int32",
            "description": "経度方向のインデックス。",
            "example": 931386,
            "minimum": 0
          },
          "yZoomlevel": {
            "type": "integer",
            "format": "int32",
            "description": "緯度方向のズームレベル。",
            "example": 20,
            "maximum": 30,
            "minimum": 0
          },
          "yIndex": {
            "type": "integer",
            "format": "int32",
            "description": "緯度方向のインデックス。",
            "example": 412905,
            "minimum": 0
          },
          "tZoomlevel": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "時間軸のズームレベル",
            "example": 25,
            "maximum": 35,
            "minimum": 0
          },
          "tIndex": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "時間軸のインデックス。",
            "example": 7,
            "minimum": 0
          }
        }
      },
...
      "InsertDataRequest": {
        "type": "object",
        "description": "空間IDの配列を指定して値を挿入する",
        "required": [
          "value",
          "spatial_ids"
        ],
        "properties": {
          "value": {},
          "spatial_ids": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpatialId"
            }
          },
          "zoom_level_policy": {
            "$ref": "#/components/schemas/ZoomLevelPolicy"
          }
        }
      },
...
      "RangeId": {
        "type": "object",
        "description": "区間表現の時空間IDを表す型。\n\n時間を指定する場合は `i` と `t` を必ず両方指定すること。\n両方省略した場合は「全時間」を表す。",
        "required": [
          "z",
          "f",
          "x",
          "y"
        ],
        "properties": {
          "z": {
            "type": "integer",
            "format": "int32",
            "description": "ズームレベル。",
            "example": 20,
            "maximum": 30,
            "minimum": 0
          },
          "f": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "description": "高さ方向の範囲 `[min, max]`（両端を含む）。",
            "example": [
              0,
              0
            ]
          },
          "x": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            },
            "description": "経度方向の範囲 `[min, max]`（両端を含む）。",
            "example": [
              931388,
              931390
            ]
          },
          "y": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            },
            "description": "緯度方向の範囲 `[min, max]`（両端を含む）。",
            "example": [
              412900,
              412907
            ]
          },
          "i": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "時間間隔（単位: 秒）。1970-01-01T00:00:00Z を起点とする。任意の値は指定できず、 次の暦の単位のいずれかのみ指定できる: `1`、`60`、`3600`、`86400`、`34359738368`（システムの全時間）。",
            "example": 3600,
            "maximum": 34359738368,
            "minimum": 1
          },
          "t": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            },
            "description": "時間インデックスの区間表現。",
            "example": [
              0,
              5
            ]
          }
        }
      },
      "RemoveDataRequest": {
        "type": "object",
        "description": "空間IDの配列を指定して値を削除する",
        "required": [
          "spatial_ids"
        ],
        "properties": {
          "spatial_ids": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpatialId"
            }
          },
          "zoom_level_policy": {
            "$ref": "#/components/schemas/ZoomLevelPolicy"
          }
        }
      },
...
      "SingleId": {
        "type": "object",
        "description": "通常の時空間IDを表す型。\n\n時間を指定する場合は `i` と `t` を必ず両方指定すること。\n両方省略した場合は「全時間」を表す。",
        "required": [
          "z",
          "f",
          "x",
          "y"
        ],
        "properties": {
          "z": {
            "type": "integer",
            "format": "int32",
            "description": "ズームレベル。",
            "example": 20,
            "maximum": 30,
            "minimum": 0
          },
          "f": {
            "type": "integer",
            "format": "int32",
            "description": "高さ方向のインデックス。",
            "example": 0
          },
          "x": {
            "type": "integer",
            "format": "int32",
            "description": "経度方向のインデックス。",
            "example": 931386,
            "minimum": 0
          },
          "y": {
            "type": "integer",
            "format": "int32",
            "description": "緯度方向のインデックス。",
            "example": 412905,
            "minimum": 0
          },
          "i": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "時間間隔（単位: 秒）。1970-01-01T00:00:00Z を起点とする。任意の値は指定できず、 次の暦の単位のいずれかのみ指定できる: `1`、`60`、`3600`、`86400`、`34359738368`（システムの全時間）。",
            "example": 3600,
            "maximum": 34359738368,
            "minimum": 1
          },
          "t": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "時間インデックス。",
            "example": 1,
            "minimum": 0
          }
        }
      },
      "SpatialId": {
        "oneOf": [
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/SingleId"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "singleId"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/RangeId"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "rangeId"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/FlexId"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "flexId"
                    ]
                  }
                }
              }
            ]
          }
        ]
      },
...
      "TableConstraints": {
        "oneOf": [
          {
            "type": "object",
            "description": "`Text` 型に対する制約。",
            "required": [
              "type"
            ],
            "properties": {
              "min_length": {
                "type": [
                  "integer",
                  "null"
                ],
                "description": "最小文字数",
                "minimum": 0
              },
              "max_length": {
                "type": [
                  "integer",
                  "null"
                ],
                "description": "最大文字数",
                "minimum": 0
              },
              "type": {
                "type": "string",
                "enum": [
                  "Text"
                ]
              }
            },
            "example": {
              "max_length": 100,
              "min_length": 1,
              "type": "Text"
            }
          },
          {
            "type": "object",
            "description": "`Int` 型に対する制約。",
            "required": [
              "type"
            ],
            "properties": {
              "min": {
                "type": [
                  "integer",
                  "null"
                ],
                "format": "int64",
                "description": "最小値（境界値を含む）"
              },
              "max": {
                "type": [
                  "integer",
                  "null"
                ],
                "format": "int64",
                "description": "最大値（境界値を含む）"
              },
              "type": {
                "type": "string",
                "enum": [
                  "Int"
                ]
              }
            },
            "example": {
              "max": 100,
              "min": 0,
              "type": "Int"
            }
          },
          {
            "type": "object",
            "description": "`Float` 型に対する制約。",
            "required": [
              "type"
            ],
            "properties": {
              "min": {
                "type": [
                  "number",
                  "null"
                ],
                "format": "double",
                "description": "最小値（境界値を含む）"
              },
              "max": {
                "type": [
                  "number",
                  "null"
                ],
                "format": "double",
                "description": "最大値（境界値を含む）"
              },
              "type": {
                "type": "string",
                "enum": [
                  "Float"
                ]
              }
            },
            "example": {
              "max": 100.0,
              "min": 0.0,
              "type": "Float"
            }
          },
          {
            "type": "object",
            "description": "`Enum` 型に対する制約。",
            "required": [
              "choices",
              "type"
            ],
            "properties": {
              "choices": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "許容される文字列の配列"
              },
              "mapping": {
                "type": "object",
                "description": "内部用: 文字列から整数IDへのマッピング",
                "additionalProperties": {
                  "type": "integer",
                  "format": "int32",
                  "minimum": 0
                },
                "propertyNames": {
                  "type": "string"
                }
              },
              "next_id": {
                "type": "integer",
                "format": "int32",
                "description": "内部用: 次に割り当てる整数ID",
                "minimum": 0
              },
              "type": {
                "type": "string",
                "enum": [
                  "Enum"
                ]
              }
            },
            "example": {
              "choices": [
                "red",
                "blue",
                "green"
              ],
              "type": "Enum"
            }
          }
        ],
        "description": "テーブルの値に対する制約。\n\n`type` フィールドで制約の種類を指定する。指定できる種類は対象のデータ型に依存する。\n- `Text`: 文字列の長さ制約（`min_length`, `max_length`）\n- `Int` / `Float`: 数値の範囲制約（`min`, `max`）\n- `Enum`: 許容される選択肢の制約（`choices`）\n\n`Presence` および `Boolean` には制約を指定できない。"
      },
      "TableDataType": {
        "type": "string",
        "description": "Table内の時空間IDに付与する値の型。",
        "enum": [
          "Text",
          "Int",
          "Float",
          "Boolean",
          "Enum",
          "Presence"
        ]
      },
      "TableInfoResponse": {
        "type": "object",
        "description": "単一テーブルの詳細情報。保持する空間ID(FlexId)の総数 `count` を必ず含む。",
        "required": [
          "name",
          "data_type",
          "max_zoom_level",
          "count"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "example_table"
          },
          "data_type": {
            "$ref": "#/components/schemas/TableDataType"
          },
          "max_zoom_level": {
            "type": "integer",
            "format": "int32",
            "example": 25,
            "minimum": 0
          },
          "count": {
            "type": "integer",
            "format": "int64",
            "example": 100,
            "minimum": 0
          },
          "constraints": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TableConstraints"
              }
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "example": "テーブルの説明文です"
          }
        }
      },
...
      "TableSummary": {
        "type": "object",
        "description": "テーブルの基本情報。一覧で返す軽量ビューで、件数は含まない。",
        "required": [
          "name",
          "data_type",
          "max_zoom_level"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "example_table"
          },
          "data_type": {
            "$ref": "#/components/schemas/TableDataType"
          },
          "max_zoom_level": {
            "type": "integer",
            "format": "int32",
            "example": 25,
            "minimum": 0
          },
          "constraints": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TableConstraints"
              }
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "example": "テーブルの説明文です"
          }
        }
      },
...
      "UpdateTableConstraints": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "type"
            ],
            "properties": {
              "min_length": {
                "type": [
                  "integer",
                  "null"
                ],
                "minimum": 0
              },
              "max_length": {
                "type": [
                  "integer",
                  "null"
                ],
                "minimum": 0
              },
              "type": {
                "type": "string",
                "enum": [
                  "Text"
                ]
              }
            },
            "example": {
              "max_length": 100,
              "min_length": 1,
              "type": "Text"
            }
          },
          {
            "type": "object",
            "required": [
              "type"
            ],
            "properties": {
              "min": {
                "type": [
                  "integer",
                  "null"
                ],
                "format": "int64"
              },
              "max": {
                "type": [
                  "integer",
                  "null"
                ],
                "format": "int64"
              },
              "type": {
                "type": "string",
                "enum": [
                  "Int"
                ]
              }
            },
            "example": {
              "max": 100,
              "min": 0,
              "type": "Int"
            }
          },
          {
            "type": "object",
            "required": [
              "type"
            ],
            "properties": {
              "min": {
                "type": [
                  "number",
                  "null"
                ],
                "format": "double"
              },
              "max": {
                "type": [
                  "number",
                  "null"
                ],
                "format": "double"
              },
              "type": {
                "type": "string",
                "enum": [
                  "Float"
                ]
              }
            },
            "example": {
              "max": 100.0,
              "min": 0.0,
              "type": "Float"
            }
          },
          {
            "type": "object",
            "required": [
              "type"
            ],
            "properties": {
              "choices": {
                "type": [
                  "array",
                  "null"
                ],
                "items": {
                  "type": "string"
                }
              },
              "add_choices": {
                "type": [
                  "array",
                  "null"
                ],
                "items": {
                  "type": "string"
                }
              },
              "remove_choices": {
                "type": [
                  "array",
                  "null"
                ],
                "items": {
                  "type": "string"
                }
              },
              "type": {
                "type": "string",
                "enum": [
                  "Enum"
                ]
              }
            },
            "example": {
              "add_choices": [
                "yellow"
              ],
              "remove_choices": [
                "red"
              ],
              "type": "Enum"
            }
          }
        ]
      },
      "UpdateTableRequest": {
        "type": "object",
        "description": "テーブルの名前変更または制約を更新する",
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ],
            "example": "example_table_renamed"
          },
          "constraints": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/UpdateTableConstraints",
                "description": "更新後の値に対する制約。指定しない場合は制約を削除する。"
              }
            ]
          },
          "validate_existing_data": {
            "type": "boolean",
            "description": "既存データが新しい制約を満たしているか検証するかどうか。\n\n- `true` (デフォルト): 既存データが新しい制約に違反している場合、更新を却下してエラー (400) を返します。\n- `false`: 既存データの走査をスキップして即座に変更を適用します。高速に変更できますが、制約に適合しない不整合なデータが残るリスクがあります。",
            "default": true,
            "example": true
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "example": "更新されたテーブルの説明文です"
          }
        }
      },
...
      "ZoomLevelPolicy": {
        "type": "string",
        "description": "テーブルの max_zoom_level よりも大きなズームレベル値の空間IDが入力された場合の処理方針",
        "enum": [
          "Error",
          "Ignore",
          "Normalize"
        ]
      }
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}